requirements.txt
replicate
python-dotenv
pandas
typer
rich
